flow: add a flow lock for packet threads
Flag introduces a lock to the Flow objects that makes the Flow object
unavailable to packet threads but still available to management threads.
When a new packet of the flow is received, the packet threads should,
instead of using the existing one, create a new flow record. The lock
is expected to be used for possibly soon-to-be-evicted flows where
touching the existing flow would not be correct.